Move `Redirection` from `core-utils` to `common-utils`
Slightly modify the code so it can be used independently of ORT.
